Feature/add sound

+++ ev3dev2simulator/connection/message/SoundCommand.py
@@ -3,7 +3,7 @@ class SoundCommand:
     Command send from the ev3dev2 mock to the simulator telling the robot to display the supplied message.
     """
 
-    def __init__(self, message: str, duration: int, sound_type: str):
+    def __init__(self, message: str, duration: float, sound_type: str):
         self.message = message
         self.duration = duration
         self.sound_type = sound_type
